feat: add client.connect method and always pass in account and ledger for PaymentRequest
the client needs to be connected in order for a PaymentRequest to use the client's account and
ledger details. previously, packets were being generated with a null account and ledger because the
client was not yet connected.

client.connect must now be called before client.createRequest and
PaymentRequest requires account and ledger to be explicitly passed in