feat: add roundingMode
either the receiver or an individual request can be created with a
roundingMode, such that amounts with too many decimal places will be
rounded accordingly (request param takses precedence). it accepts all of BigNumber.js' rounding modes
https://mikemcl.github.io/bignumber.js/#constructor-properties. by
default it will not round request amounts and amounts with too many
decimal places will cause an error to be thrown. errors will also be
thrown if rounding a very small amount would reduce it to zero or more
than double it.

resolves https://github.com/interledger/js-ilp/issues/17