fix: handle trailing zeros in packet amount
integration tests were failing because the packet amount was changed somewhere along the way to

include trailing zeros. because of this, the receiver thought the packet didn't match the original

condition. by using the bignumber.js .toString method in the initial packet creation and the

re-generation, we ensure that it will handle trailing zeros the same way in both cases