fix: don't reject when condition doesn't match
cases where an incoming transfer's condition doesn't match what the receiver would generate could be
caused by someone tampering with the packet OR by the request being created by another receiver
listening on the same account. therefore, we shouldn't cancel incoming transfers just because the
condition doesn't match what we would generate, in case the request was created by another receiver.
note that this could be addressed in two other ways as well: adding a receiver-specific uuid to the
request address or adding a separate hash of the request details as an explicit checksum. by adding
a receiver-specific id to the request address, the receiver would know if the payment was destined
for it or not, however this would come with a potential loss of privacy because this would allow
others to correlate which payments are for a specific receiver (presumably tied to one application)
on an account. the other option of adding an explicit checksum would allow us to reject incoming
transfers that don't match the checksum, but if the packet is deliberately tampered with, the
attacker could trivially change the checksumso it seems unlikely that adding the extra data would
provide much benefit since most transfers with maliciously altered packets would have to wait until
they expire anyway.

fixes https://github.com/interledger/js-ilp/issues/43

chore: update integration tests