fix: make PSK actually work
the PSK functionality was not being tested to see that it would actually fulfill payments. it turned out that it wouldn't because the sender and recipient were not encrypting and MACing in the same order. this also moves groups the PSK tests together so it's easier to see what behavior is being tested